backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
python-dotenv>=1.0.1

# Optional (only used if ANTHROPIC_API_KEY is set)
anthropic>=0.40.0

